fix factory tests so they don't break when new items are registered
Fix the factory tests for rules and actions so that they don't break
if we add new items. They test a static list of registered names now,
but test each one separately and don't require the list of registered
names to match exactly the values being tested, so we can have extra
items registered.

Closes bug #24