client: update ssl handling for newer python 3 and imapclient
The imapclient function for creating a default SSL context has been
removed, so replace its use in the client module with the function
from the ssl module in stdlib instead.

Update the way the cafile and check_hostname settings are managed
based on examples in the ssl module documentation.

Update the mocks in the related unit test to match the new behavior.