fix SortByYear tests
Using __setitem__ with the email class appends to the existing header,
so to replace the header it is necessary to first delete it. Also add
some test cases for variations of unparsable dates, and then fix the
tests so the right mailbox is used when the date can't be parsed.